requirements.txt
fastapi>=0.128.8
uvicorn>=0.39.0
anthropic>=0.86.0
slack_sdk>=3.41.0
jira>=3.8.0
opsgenie-sdk>=2.1.5
boto3>=1.42.74
pytest>=8.4.2
PyGithub>=2.9.0
kubernetes>=35.0.0
groq>=1.0.0
python-dotenv>=1.2.1
langgraph>=0.2.0
openai>=1.30.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.9
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
tenacity>=8.2.0
pybreaker>=1.0.1
prometheus_client>=0.20.0
